split code to support freebsd
